requirements.txt
# Core Dependencies (will be expanded later)
python-dotenv==1.0.1
requests==2.32.3
openai>=1.0.0
schedule==1.2.1
Pillow>=10.0.0
# System monitoring
psutil>=5.9.0
# Development/Validation Dependencies
pytest>=7.4.0
flake8>=6.1.0
mypy>=1.5.0
bandit>=1.7.5

# Web GUI
streamlit>=1.36.0
streamlit-option-menu>=0.3.6
watchdog>=3.0.0
